requirements.txt
Flask
requests
beautifulsoup4
fpdf2
argparse
